feat(funding): combine API WISH-129
